+++ app/layout.tsx
@@ -27,9 +27,9 @@ export default function RootLayout({
   return (
     <html lang="en" className='scroll-smooth'>
       <body className={cn(
-        "bg-gray-50 text-gray-950 relative font-sans", fontSans.variable)}>
-        <div className='bg-[#dbd7fb] absolute top-[-6rem] right-[11rem] h-[31.23rem] w-[31.23rem] rounded-full blur-[10rem] sm:w-[68.75rem] -z-10'></div>
-        <div className='bg-[#fbe2e3] absolute top-[-1rem] left-[-35rem] h-[31.23rem] w-[31.23rem] rounded-full blur-[10rem] sm:w-[68.75rem] md:left-[-33rem] lg:left-[-28rem] xl:left-[-15rem] 2xl:left-[-5rem] -z-10'></div>
+        "bg-gray-50 dark:bg-[#0A192E] text-gray-950 dark:text-gray-100 relative font-sans", fontSans.variable)}>
+        <div className='bg-[#dbd7fb] dark:bg-black absolute top-[-6rem] right-[11rem] h-[31.23rem] w-[31.23rem] rounded-full blur-[10rem] sm:w-[68.75rem] -z-10'></div>
+        <div className='bg-[#fbe2e3] dark:bg-black absolute top-[-1rem] left-[-35rem] h-[31.23rem] w-[31.23rem] rounded-full blur-[10rem] sm:w-[68.75rem] md:left-[-33rem] lg:left-[-28rem] xl:left-[-15rem] 2xl:left-[-5rem] -z-10'></div>
         <ThemeProvider attribute='class' defaultTheme='system' enableSystem disableTransitionOnChange>
           <ActiveSectionContextProvider>
             <Header />