+++ components/contact.tsx
@@ -58,17 +58,17 @@ const ContactComponent = () => {
     return (
         <section ref={ref} id='contact' className='lg:max-w-[50rem] xxs:w-64 xs:max-w-sm sm:max-w-sm sm:mb-0 md:mb-28 scroll-mt-28 mt-4 xxs:mt-10'>
             <SectionHeading heading={'Contact Me'} />
-            <p className='text-gray-800 text-lg'>Currently I&apos;m actively looking for any new opportunities, my inbox is always open. Whether you have a question or just want to say hi, I&apos;ll try my best to get back to you!</p>
+            <p className='text-gray-800 dark:text-gray-400 text-lg'>Currently I&apos;m actively looking for any new opportunities, my inbox is always open. Whether you have a question or just want to say hi, I&apos;ll try my best to get back to you!</p>
             <Sheet>
                 <SheetTrigger asChild>
                     <span className='mt-10 flex flex-row justify-center items-center'>
                         <StylishButton />
                     </span>
                 </SheetTrigger>
-                <SheetContent side={'right'} className='bg-gradient-to-r from-blue-100 to-yellow-100'>
+                <SheetContent side={'right'} className='bg-gradient-to-r from-blue-100 to-yellow-100 dark:bg-gradient-to-r dark:from-sky-900 dark:to-gray-800'>
                     <SheetHeader>
-                        <SheetTitle className='text-xl'>Hi There!</SheetTitle>
-                        <SheetDescription className='text-lg'>
+                        <SheetTitle className='text-xl dark:text-gray-200'>Hi There!</SheetTitle>
+                        <SheetDescription className='text-lg text-gray-400'>
                             You can get in touch with me or follow me on these channels.
                         </SheetDescription>
                     </SheetHeader>